+++ CHANGELOG.md
@@ -15,6 +15,10 @@
   - Feat: add typed error.
   - Feat: add `ILexer.takeUntil`. [#16](https://github.com/DiscreteTom/retsac/issues/16)
   - Feat: add `Lexer.utils.regexLiteral`. [#15](https://github.com/DiscreteTom/retsac/issues/15)
+- Parser
+  - **_Breaking Change_**: add `name` for `ASTNode` and `ASTObj`.
+  - **_Breaking Change_**: `ASTNode.toTreeString` will also print `ASTNode.name`.
+  - **_Breaking Change_**: refactor `ASTNode.toString`. The output format is changed.
 - Optimize performance.
   - [#6](https://github.com/DiscreteTom/retsac/issues/6)
   - Lexer will skip unexpected actions by checking expected text.